pyproject.toml
[tool.poetry]
name = "simple_social_media"
version = "0.1.0"
description = ""
authors = ["Daniil Marichuk"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.6"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.1"}
mysql-connector-python = "^8.0.29"
requests = "^2.28.1"
Faker = "^13.14.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
